requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
flask-cors==4.0.0
python-dotenv==1.0.1

google-cloud-storage==2.19.0
google-cloud-tasks==2.19.3
google-cloud-documentai==3.5.0
google-cloud-aiplatform==1.108.0

psycopg2-binary==2.9.9
pgvector==0.2.5

markitdown[all]==0.1.2

